requirements.txt
Django==1.9.1
MySQL-python==1.2.5
Pillow==3.1.0
PyJWT==1.4.0
PyYAML==3.11
boto==2.39.0
boto3==1.2.3
botocore==1.3.23
django-braces==1.8.1
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-filter==0.12.0
django-oauth-toolkit==0.10.0
django-registration==2.0.4
django-rest-swagger==0.3.5
django-storages==1.1.8
django-suit==0.2.17
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
docutils==0.12
futures==3.0.4
gunicorn==19.1.1
jmespath==0.9.0
meld3==1.0.0
oauthlib==1.0.3
python-dateutil==2.4.2
six==1.10.0
sqlparse==0.1.18
wsgiref==0.1.2
